python/requirements.txt
pandas
scikit-learn==1.5.0
joblib
paho-mqtt>=2.0,<3.0
influxdb-client>=1.34.0
